An improved method-table based object system.
Moved dispatcher into its own module.  The objects module depends
on the maps module, which in turn depends on the dispatcher module.

Each memory-allocated object's first field points to its type.
The type's third field points to the method table.  The method table
maps selectors (symbols) to procedures whose first argument is the
receiver.

Method definition
-----------------

Using defm, new methods can be added to existing types.  For instance,

   (defm @smallint 'plus (n)
     (+ self n))

defines the method 'plus for small integers.  The method takes one
parameter which is added to the receiver and returned.  The macro
creates the function

   (lambda (self n)
     (+ self n))

and registers it in the method table.  Note that the self argument is
implicit.

Message sending
---------------

To send a message to an object, the send method is used:

   (send 4 'plus 3)  ; --> 7

This method may be renamed in the future to facilitate left-to-right
reading.