Compile all globally visible procedures on REPL startup.
This also fixes a bug in the compiler where the expressions in a
`set!' and a `def' got called with a tail-call.  (This is of course
not allowed, since it misses to do the assignment.)  On the old
version, the behaviour could be reproduced by doing

  (c! assemble)
  (c! cons)
  (c! someotherprocedure)  ; result will be broken.

The `assemble' procedure has some (set! ...) forms in tail call
positions (hidden behind the `add-to-list!' macro).  These only get
executed as tail calls when the executed procedure (`cons') is also
byte-compiled, so that the tail call becomes possible.  In that case,
the assignment was missed and the postponed actions (see code for
`assemble') were not executed later, resulting in bad results for
future compilations.

Also,
 - more debugging output in debug.c
 - proper stack traces within the bytecode interpreter
 - make array functions available from C
 - increase some memory size limits as needed
 - compile the whole system on REPL startup
 - escaping bug in array construction `list->array'
 - bounds checking for local variable accesses from bc interpreter
   (only if debugging mode is on)