Tail call optimization for the C-based expression evaluator.
Currently evalutated functions are not properly tracked yet.

Some of the examples programs were too big for the C stack when trying
to load them with the C-based expression evaluator.  This is of course
due to the PEG framework using continuation-passing style, so that
parsing a bigger input will add a few stack frames for each character
in the input, and the stack size is easily exhausted if you don't have
tail call optimization.  These examples work now.