Make stack access indirect via a pointer.
This uncovered a bug in the interpreter:
We used to call interpret() without pushing anything to the "apply
stack" (see procedures.h).  This doesn't work, because we pop the
apply stack on tail calls.  When doing a tail call, we first pop
something, and then write a new value to the wrong place.  In this
case, moving the Lisp stack to the C heap moved it next to the apply
stack.  When the apply stack was misused, it would corrupt the `size'
field of the Lisp stack.

Running the tests also uncovered a bug in the continuations-test.fn,
where a function had one argument too much.