Make interpreter_state_t be a "cache" for the current frame.
This makes retptrs on the stack obsolete, as returning can now be
done through frames, which know about the caller frame.

In this change, we still have both the retptr structure and the
frames, and some debugging code that makes sure they are in sync
during execution.  This helped to migrate.