A bit of test file caching.
This is only caching at the Lisp-expression level,
it's not caching the compiled code.

You can't just cache compiled code easily, because the execution of
one expression in a file can have side-effects on the compile result
for the next, and so we probably need to take this state into account
for the lookup.