A bytecode compiler written in C.
This doesn't deprecate the compiler written in Lisp.
It would be much nicer to translate the Lisp compiler to C
or to simply make it run much faster.

This allows to delete most of eval.c and code related to the overly
complicated bootstrapping, and speeds up bootstrapping considerably
too.