read-field, write-field bytecodes and support in the compiler.
These bytecodes allow to access object fields by index directly without calling
$mem-get and $mem-set!.  The drawback is that this is an unsafe operation which
is allowed from the bytecode.  On the other hand, there are already unsafe
operations that you can do with the bytecode, like calling pop too often, which
will mess up the memory.  We have to trust the compiler to some extent.

The compilation strategy is to transform $mem-get and $mem-set! calls to
bytecodes, as long as the accessed index is a fixed number.  Otherwise, we
continue to call the $mem-get and $mem-set! functions.