Make the end marker a statically defined value.
This way, the end marker can't change its location any more
and there is no way anything in the GC module can move it around.
This happened not during regular GCs, but when the full object space
got deserialized from a file. In that case, the end marker was
not updated and was pointing to garbage data.

On ARM, this lead to a crash because of broken GC invariants further
down the road.

This change also adds additional safety checks in the GC.